Add numbers file for state, improve usability
Now maintains previously added charge numbers in the file
$HOME/.clockme/numbers to persist between days. Improves the interface;
now choosing a charge number is done based on vertical order, not charge
number, for simpler entry. Status of each number is printed when a
listing is requested. Fixes a bug that underreports time if a certain
pause sequence is used.